Fix issue where display area does not take up full height
See vega/vega-lite-ui#73 and vega/polestar#322

This was introduced in dea6557d32c7733d220bd5dfca7bcc62f22b4dc8
while addressing an issue that prevented the bookmarks from closing:
the display height issue arose due to the introduction of another
layer of nesting that the CSS did not account for.

The same angular hierarchy can be achieved with less DOM nesting, so
this commit addresses the overflow by removing unneeded intermediate
div nodes.

Note that this did not end up being an issue with vega-lite-ui, so
this closes vega/vega-lite-ui#73 as a dupe of an issue that exists
separately within polestar and voyager.

Polestar commit to follow.