Re-render the visualization list when dataset changes
This closes #176, where the visualization list would not automatically
re-render until you mouseover a pill. The reason this was happening
was that the update method is debounced, and debounced methods execute
outside of Angular's regular digest cycle. As such, while the Visrec
object was updated, the change would not take effect until a mouseover
or some other angular event handler fires, triggering a new $digest.

We could solve this problem by manually `scope.$digest`ing, but this
is the use-case `$apply` was intended for: wrapping the update method
in `$apply` is the most concise solution to the problem.

![toggle](https://cloud.githubusercontent.com/assets/442115/10500061/42b770de-72a3-11e5-8bd2-58e60f04f228.gif)