Workaround for tooltip opacity issue (fixes #200)
The CSS intended to hide a tooltip when it is not open does not
work due to a CSS selector precedence issue in the angular-tooltip
code. Using `!important` is generally not best-practice, but it
does let us temporarily work around the problem described in
issue #200 and 720kb/angular-tooltips#95

Once this issue is resolved in the angular-tooltips library this
code can be deprecated and removed.