Use a dot & fix the debug button broken by previous commit
Using `ng-if` introduced a new scope, breaking the `ng-click` on the
debug button: shifting this to a function on the parent scope restores
the correct behavior.

The "About" button has no function so the `ng-click` (which just
referenced a property, without executing any code) was emptied to
emphasize the TODO functionality.