Get style source maps working (fixes #281)
In order to get source maps working for SASS, had to switch to the
`source-map` devtool, which is [recommended by `extract-text`](https://github.com/webpack-contrib/extract-text-webpack-plugin/tree/58dd5d31735dd801cba43c87700f4d6144203c75#usage).

I also had to stop using the `-d` CLI flag for `webpack`, because it
overrides `devtool` and replaced it with `--debug`, which it includes.